reporting:  REPORT-135, REPORT-290, REPORT-244, REPORT-250.  This includes the following features/changes: * Removing the current scheduled tasks and re-implementing these using non-user-configurable timer tasks managed by spring. * Making all report evaluations happen via an asynchronous process, queueing these reports, and re-directing the user to a page that has more information as to the evaluation status * Making it possible to schedule a report via a cron expression, rather than just at a regular interval * Making it possible to associate a report processor via the run/schedule report UI
These features will be part of the 0.7.1 release.

+++ web/module/reports/reportHistory.jsp
@@ -110,7 +110,7 @@
 					</c:choose>
 				</td>				
 				<td>
-					<a href="reportHistoryDelete.form?uuid=${r.uuid}"><img src='<c:url value="/images/trash.gif"/>' border="0"/></a>
+					<a onclick="return confirm('<spring:message code="reporting.reportHistory.confirmDelete"/>');" href="reportHistoryDelete.form?uuid=${r.uuid}"><img src='<c:url value="/images/trash.gif"/>' border="0"/></a>
 				</td>
 			</tr>
 		</c:forEach>
